create table row/column abstraction
